requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.2
cycler==0.12.1
datasets==4.0.0
diffusers==0.34.0
dill==0.3.8
einops==0.8.1
filelock==3.13.1
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2024.6.1
grpcio==1.74.0
hf-xet==1.1.8
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
Jinja2==3.1.4
kiwisolver==1.4.9
lazy_loader==0.4
lpips==0.1.4
Markdown==3.8.2
MarkupSafe==2.1.5
matplotlib==3.10.5
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
packaging==25.0
pandas==2.3.2
pillow==11.0.0
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
pyarrow==21.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.15.3
six==1.17.0
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tifffile==2025.5.10
torch==2.5.1+cu121
torch-fidelity==0.3.0
torchvision==0.20.1+cu121
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
